[POC] Featured Image
- address feedback from Burkhard on PR
